requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
python-dotenv==1.0.0
openai==0.28.0
